+++ src/main/java/racingcar/racing/Cars.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.racing;
 
 import java.util.List;
 import java.util.stream.Collectors;
