+++ src/main/java/racingcar/racing/RacingCar.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.racing;
 
 public class RacingCar {
     private Cars cars;
